Implement touchscreen (mouse + optional right analog)
A virtual cursor is drawn on the lower screen to represent the emulated cursor. Left click using the mouse, or L3 to touch the screen.

Because the right analog is also shared with the C-Stick, a few options are provided:
 - Because so few games actually use it, you can continue with both C-Stick and touchscreen emulation.
 - You can disable the C-Stick entirely, or
 - You can disable the joypad's emulated pointer and use it exclusively for the C-Stick.
This is toggleable in the settings.

Acceleration is hardcoded at the moment, and L3 potentially conflicts with LR (though, again, very few games use them).